create default rule manager (#467)
Create a default rule manager implementation to be used between all
engines to load rules and text units. Previously each rule implements
its own version of RuleManager to load the rules and text units, with
common steps between them, just changing the GetRules implementation to
return the rules of specific language, with this new default
implementation we can just pass this rules to constructor and use the
same logic to all engines.