cli: fix --ignore-severity flag
Previously when we use the flag --ignore-severity to ignore kinds of
vulnerabilities based on severity we still saw the vulnerabilities.
This commit fix this behaviour to ignore all vulnerabilities listed
on ignore-severity flag based on their severity.
This commit also fix and update the test case to validate this feature.

Fixes: #526

Co-authored-by: Matheus Alcantara <matheus.alcantara@zup.com.br>