cli: fix unnecessary log file creation
Previously we are reading and evaluating the config values on every
command of CLI, even if we don't needed. For example, when we ran the
version command we create the log files. So this commit change to read
and merge config values and create the log file only when needed.

Co-authored-by: Matheus Alcantara <matheus.alcantara@zup.com.br>