config: fix flags not overriding config file and env variables
Previously we are parsing the config values in follow order:
command line flags -> config file -> env variables.
This commit fix the order to parse in follow order:
config file -> env variables -> command line flags.
Since the func passed to PreRun field on cobra command is executed after
the command line parsing, a change was made on CreateStartCommand method
to call the config methods to load the values from config file and
environment variables before creating the cobra command, so when the
PreRun is execute the only thing that is needed to do is normalize the
values and setup the log configuration.

Signed-off-by: Matheus Alcantara <matheus.alcantara@zup.com.br>