usecases: remove unnecessary partial struct to validate config
Previously to validate the user input config we create a new struct
`ConfigToValidate` with all the fields of config and them validate,
since the fields from config is all public, this commit remove the
struct `ConfigToValidate` and check all fields from Config struct
directly. An improvement on error message also was made to make it
clearer.

Signed-off-by: Matheus Alcantara <matheus.alcantara@zup.com.br>