config: fix read config file if its not default value
Previously we are reading the config file before parsing the command
line flags, so, when a user pass a different path of config
file we weren't read the file because we was trying to read on default
value (that is the current working directory). This commit change to
parse the values manually, so, we first parse global flags, them read the
config file and environment variables and finally parse the start command
flags. The changes on config/config.go is basically a rollback from changes
made on pr #523.

Issue: #656

Signed-off-by: Matheus Alcantara <matheus.alcantara@zup.com.br>