git: improvement on success commit author tests
Previously on tests of success getting commit author we was just
asserting that the result is not empty, but actually the tests was not
getting the correct commit author and was returning commit author not
found. This commit add asserts to make sure that we are successful
getting the commit author info from a file and line.

Signed-off-by: Matheus Alcantara <matheus.alcantara@zup.com.br>