tests: create package testutil
Previously for each test case that need to use the examples directory is
necessary that theses tests use `filepath.Join` to get the correct path
to these examples. This commit introduce a new package `testutil` to
centralize these paths and other test helpers. Note that this commit
only add the `testutil` package, we still need update the test cases to
use this package.

Signed-off-by: Matheus Alcantara <matheus.alcantara@zup.com.br>