git: validate if repository is fully cloned
Previously if commit author was enable and the repository was cloned
using --depth flag the results of commit author was wrong since not
all commits are present in the repository. When a repository is cloned
using --depth flag a file named `shallow` is created inside .git
directory, so this commit add a validation to check if this file exists
when starting analysis, if this file exists a warning is printed
informing the user that the commit authors results could be wrong.

Signed-off-by: Matheus Alcantara <matheus.alcantara@zup.com.br>