Adding tests for rules 7-17 of dart language
Note that in the HS-DART-8 rule, I increased the accuracy of the rule to only get logs that have string concatenation or string format.
Note too the HS-DART-15 rule was not working because there is no capitalized property so I added the option to include it in the entire regular expression.
Signed-off-by: wilian <wilian.silva@zup.com.br>