utils/file: remove usage of find command to search for files
Previously some functions of utils/file package use the command find
to search for files, which make only Linux compatible and add an
unnecessary dependency. This commit change the implementation to use
filepath.Walk function and search using only the Go std.

This commit also make some improvements on function names, remove
dead code and add some documentation.

Also this commit change some usages of ReplacePathSeparator to use
filepath.Join instead which is a better approach since this package is
OS compatible. Note that not all usages of ReplacePathSeparator was
changed since in some cases we need to replace slashes on paths that
was returned from tools that run on Docker when we are running on
Windows.

Signed-off-by: Matheus Alcantara <matheus.alcantara@zup.com.br>