workflow: execute unit tests on Windows
Previously some unit tests was implemented to execute and validate
scenarios only for Linux based machines. This commit fix these tests to
execute on Windows.

Note that the goal of this commit is only to fix these tests and make
them work on Windows, we still need to make some improvements on some
of these tests to have a better assertiveness.

This commit also change the configuration of Test workflow to add
windows-latest as a new runner.

Signed-off-by: Matheus Alcantara <matheus.alcantara@zup.com.br>