workflow: execute unit tests on MacOS
Previously the configuration of workflow tests was testing only Linux
and Windows. This commit add macOS-latest as a new runner.

Note that the goal of this commit is only to add macOS runner and fix
some tests to make them work on MacOS, we still need to make some
improvements on some of these tests to have a better assertiveness.

Signed-off-by: Matheus Alcantara <matheus.alcantara@zup.com.br>