docker: improvement logs on pulling image
Previously when we call `checkIfImageNotExists` function we check for
errors and if image does not exists on the same statement and only log
if some errors occurs, which turn the code a bit confuse to read.

This commit change to check for errors and if image does not exists on
different statements and also add a debug logging to inform that image
does not exists on cache and we will try to pull from registry.

This commit also add some tests to assert that we are pulling the image
only if its not exists on cache.

Signed-off-by: Matheus Alcantara <matheus.alcantara@zup.com.br>