analyzer: disable scan loading when debug level log is set
Previously when we set the log level for debug the output is hard to
read since the we are printing the debug messages and the scan loading.

This commit change the newScanLoading constructor to set the writer of
loading to io.Discard when the log level is debug or trace.

Signed-off-by: Matheus Alcantara <matheus.alcantara@zup.com.br>