printresults: add tests to cover output and Sonarqube output type (#760)
This commit add new tests to cover Sonarqube output type and add asserts
to check if what was printed is correctly.

The tests was changed to use table testings approach to make more easily
to add a new testcase.

The PrintResults implementation was improved too. Basically a new
io.Writer field was added to customize where we will write outputs. The
default constructor will always write to Stdout, but on tests we use a
custom BufferString to write.

This commit also make some improvements on code organization and private
method names.

Updates #718

Signed-off-by: Matheus Alcantara <matheus.alcantara@zup.com.br>