formatters: fix random vulnerability hash on Windows
Previously when running the Horusec on Windows we was getting a
different vulnerability hash for each analysis, this problem was
happening because we weren't removing the .horusec folder from
the file path before generating the hash, and since the .horusec folder
contains a uuid, on every analysis we was getting a new uuid and
consequently a new vulnerability hash.

The commit fix this issue using a better approach to remove the .horusec
folder from filepath, using the relative path from file using the
.horusec folder as base path.

This commit also change the testcase to catch this bug.

This bug was founded on #777

Signed-off-by: Matheus Alcantara <matheus.alcantara@zup.com.br>