workdir:chore - improve tests and rename public functions
This commit improve the current tests and add new ones. The improvement
was made to have a better assertiveness and cover more scenarios.

The public functions `NewWorkDir` and `ParseInterfaceToStruct` was
renamed to `Default` and `MustParseWorkDir` respectively to follow the
same API from toolsconfig and customimages packages to have consistency
between them.

This commit also remove the unused method `Type` from WorkDir struct.

Updates #718

Signed-off-by: Matheus Alcantara <matheus.alcantara@zup.com.br>