gosec:chore - improve tests and clean code
Previously the test cases of Gosec formatter was just testing that the
`StartAnalysis` method was not panicking and does not do any assert on
the results of analysis.

This commit improve the tests to assert that the results was parsed and
add correctly on Analysis.

This commit also make some clean up on code removing the unnecessary
package `entities` and move the schema of gosec output to gosec package
properly. Since these output schemas is not used by any package they was
changed to be private too.

Updates #718

Signed-off-by: Matheus Alcantara <matheus.alcantara@zup.com.br>