nancy:chore - improve tests and code cleaning
This commit improve the test cases of parsing nancy results to verify
that all fields was filled correctly to all platforms.

This commit also move the entities and enums from nancy/entities and
nancy/enums to nancy package and make them private.

Some clean code and code organization was also made to improve
readability.

This commit also fix a bug on `file.GetDependencyCodeFilepathAndLine` that
previously we was not closing the file, which causes we cannot remove the
.horusec folder in Windows.

Updates #718

Signed-off-by: Matheus Alcantara <matheus.alcantara@zup.com.br>