dotnetcli:chore - improve tests asserts and code cleaning
Previously the mock output from success tests was a bit hard to read and
understand, and we was not asserting that the formatter was parsing
correctly the output from dotnet cli.

Since the formatter needs a .csproj file to check the dependencies the
function created on #806 to copy files from examples directory to a
.horusec directory was moved to testutil package (and changed to be more
generic) and the references on nancy tests was updated.

This commit improve the tests to use .csproj files from examples
directory and new asserts was added to verify that the output was
parsed correctly.

The schema definition and enums was also moved from dotnet_cli/entities
and dotnet_cli/enums to dotnet_cli package and changed to be private.

Since the const `CsProjExt` is used on tests of `utils/file` package the
package name of these tests was renamed to `file_test` to avoid
cyclomatic imports.

Updates #718

Signed-off-by: Matheus Alcantara <matheus.alcantara@zup.com.br>