flawfinder:chore - improve tests asserts and code cleaning
Improves the test cases of errors to verify that errors were actually
added to the analysis. Some asserts was also added on success tests to
verify that the parsing is correctly and all fields is filled.

This commit also move schema definition of flawfinder results from
flawfinder/entities to flawfinder package and make them private.

Finally some code organization was made and redundant code was removed.

Updates #718

Signed-off-by: Matheus Alcantara <matheus.alcantara@zup.com.br>