images:chore - upgrade Docker tags
This commit upgrade the tags for almost all languages that is supported
by Horusec.

Python was not upgraded because the safety and bandit was not released
with official support for Python 3.10.

Generic and HCL was not upgraded too because we need to upgrade the
installation script of tools that is installed on Dockerfile (tfsec,
semgrep and dependency check) which will be done separately.

Signed-off-by: Matheus Alcantara <matheus.alcantara@zup.com.br>