scs:chore - improve tests asserts and code cleaning
Previously the mock output from success tests was a bit hard to read and
understand, and we was not asserting that the formatter was parsing
correctly the output from security code scan.

This commit improve the tests to use .sln files from examples
directory and new asserts was added to verify that the output was
parsed correctly and all fields from vulnerability struct was filled.

The schema definition and enums was also moved from scs/entities and
scs/enums to scs package and changed to be private.

Updates #718

Signed-off-by: Matheus Alcantara <matheus.alcantara@zup.com.br>