nancy:fix - empty file name, code and line on vulnerability result
Previously when we execute Nany, we was searching the vulnerability only
on go.mod files, but Nancy analyse go.sum files too, so when some
vulnerability returned by Nany that is only  declared on go.sum the
vulnerability result show a File without file name (only the project
path) and the Code and Line empty.

This commit change the function `GetDependencyCodeFilepathAndLine` to
receive a list of extensions that can match on the given projectPath, so
the Nancy formatter was changed to pass .mod and .sum extensions. Since
the function signature was changed to be variadic the references was
changed too to follow the new parameter orders.

The success test case was also changed to cover this scenario.

Signed-off-by: Matheus Alcantara <matheus.alcantara@zup.com.br>