cli:chore - deprecate monitor-retry-count flag
Previously we control all execution of tools using a monitor package,
but this package contained a lot of data races that was fixed on #477
and the monitor retry count flag was no longer necessary, since this was
used to control the timeout counter.

This commit mark monitor-retry-count flag as hidden and deprecated, so
new users will not see this flag and users that is currently using this
flag on pipelines will see a warning saying to use just --analysis-timeout
flag.

Signed-off-by: Matheus Alcantara <matheus.alcantara@zup.com.br>