language_detect:fix ignore files correctly on Windows
Previously when checking which files should be ignored by analysis on
Windows the directories patterns was not working properly, this is
because we was expecting that these paths would be always Unix like
paths so when a user set a path `**/tests/**` we can not ignore the
tests directory on Windows since the path would be `**\\tests\\**`.

This commit fix this issue converting all paths and all patterns to Unix
like paths before checking if should be ignored. Now `**/tests/**` would
match both `foo/bar/tests` and `foo\\bar\\tests`.

If the user set the pattern to use Windows slashes like `**\\tests\\**`,
this will work correctly only on Windows, so it is recommend that users
should always use the Unix slashes when configuring patterns that should
be ignored by analysis.

Signed-off-by: Matheus Alcantara <matheus.alcantara@zup.com.br>