scripts:chore - remove redundant logging
Previously if the user input `latest` version to download we was logging
that the download would be the latest-rc and also the latest version.
Even though the logging in wrong, we don't need to inform twice which
version is being downloaded, so this commit remove the redundant logging
and also use the same error message that is showed when user pass an
invalid version to when user not pass any input.

Signed-off-by: Matheus Alcantara <matheus.alcantara@zup.com.br>