analyzer:chore - split analyzer implementation into runner
Previously the Analyzer struct holds the entire pipeline of an analysis,
which add a lot of complexity, since this pipeline has a lot of steps:
Detect languages, orchestrate tools by languages, send analysis to Horusec
manager and them print the results on stdout.

This commit split the Analyzer struct into a new struct `runner`.
This new struct will be responsible only to orchestrate all tools to
their respective languages. The coded implemented on runner is basically
a copy and paste from Analyzer implementation, so the behavior remains
the same.

Some documentation was also added to try do make more clear what are the
steps that Analyzer do, and what a runner means.

Signed-off-by: Matheus Alcantara <matheus.alcantara@zup.com.br>