formatters:fix - invalid work dir path on Windows
Previously when we were running on Windows we was setting a Windows path
as a work dir inside a Docker container. For example, we start the
analysis, and we try to find a path that contains go.mod files, we found
the path C:\user\some\path and we was using this path (with Windows
slashes) as a work dir on Docker, so when the command try to access this
path (cd some\path) we was getting errors.

This commit fix this issue converting any path to Unix path before
replacing the {{WORK_DIR}} of command.

Some tests of formatter was changed to expected that any path used on
{{WORK_DIR}} should be a Unix like path.

This commit also make some improvements on `utils/file` private
functions, resolving a TODO by refactoring the function to use a better
approach to get the relative directory of a file path. Some functions
was renamed to make more clear what they are doing and new debug logging
was added to inform the path that the searched files was founded.

Signed-off-by: Matheus Alcantara <matheus.alcantara@zup.com.br>