analyzer:fix - not printing information severity when enabled
Previously when use use the flag --information-severity the information
vulnerabilities was not been printing by PrintResults, this is because we
was getting a conflict between --information-severity and
--ignore-severity flags, the default values of --ignore-severity is INFO,
so when we use the flag --information-severity we was still not printing
because the INFO vulnerabilities was been ignored by --ignored-severity
flag.

This commit fix this issue by forcing the print of INFO vulnerabilities
when --information-severity is enabled.

Fixes #785

Signed-off-by: Matheus Alcantara <matheus.alcantara@zup.com.br>