brakeman:chore - improve tests asserts and code cleaning
Previously the tests was not asserting that we was correctly parsing the
output from brakeman and we was only checking that vulnerabilities was
being added on analysis without checking if their contents was
correctly.

This commit add some new checks to assert that all vulnerabilities
founded by brakeman is parsed correctly.

Updates #718

Signed-off-by: Matheus Alcantara <matheus.alcantara@zup.com.br>