brakeman:fix - search for Gemfile's before start analysis
If there was no `Gemfile` file in the current directory, Brakeman would
generate an error stating that the project to be analyzed was not a Ruby
on Rails project.

This commit fix this issue by looking for a directory path that contains
a `Gemfile` filename and them using this path as a work dir to execute
Brakeman.

Signed-off-by: Matheus Alcantara <matheus.alcantara@zup.com.br>