formatters/bundler:fix upgrade to fix crash on execution
Bundler on version 0.9.0 had a bug that make the execution crash, this
commit upgrade to version 0.9.0.1 that contains a fix for this issue.
Read more: https://github.com/rubysec/bundler-audit/issues/319

This commit also change the Dockerfile do install Git, since on 2.7.x
images is not installed by default, and is a requirement to execute
Bundler.
https://github.com/rubysec/bundler-audit#requirements

Signed-off-by: Matheus Alcantara <matheus.alcantara@zup.com.br>