npmaudit:chore - improve tests and code cleaning
This commit add some new asserts on successful parsing npm results
to verify that all fields of Vulnerability was filled.

Note that the test cases from `TestParseOutputNpm` was moved to
`TestNpmAuditParseOutput` to centralize all tests as it is done in
the other tests of the other formatters.

Some code organization was also made, and the entities packages was
removed and the npm schema output was moved to npmaudit package.

Updates #718

Signed-off-by: Matheus Alcantara <matheus.alcantara@zup.com.br>