bundler:fix - correctly parse output error
Previously when a project path does not have a Gemfile.lock file, the
Bundler return an error `Could not find "Gemfile.lock"` that was being
interpreted as a vulnerability. This was happening because the commit
https://github.com/rubysec/bundler-audit/commit/021f85f change the error
message from a generic error like "Errno::ENOENT" and "No such file or
directory" to a more detailed error `Could not find "Gemfile.lock"`.

Fixes #919

Signed-off-by: Matheus Alcantara <matheus.alcantara@zup.com.br>