trivy:bugfix - adding func to avoid hash changes in trivy formatter
Since the pull request https://github.com/ZupIT/horusec/pull/882 some
changes were made in the line and code of the trivy formatter,
and this data influences directly the hash generation.
This pr will avoid this hash change by using the same data as
before, but for the users the data will be showed with the fixes
made in the pull request 882, leading to no braking changes
and keeping the fixes.

Signed-off-by: Nathan Martins <nathan.martins@zup.com.br>