hash:feature - new hash format and handle the future depreciations (#996)
Currently cli generates very volatile hashes, any commit, description
changes will change the hash, wich is a problem to the vulnerabilities
management. This pull request changes the way that we generate the hash
to use only the code, line and file, this way we expect that the hash
will be more estable. In this pull request we also added a tratative to
handle with the future depreciation of the old hashes that we still
consider valid. So all users should update the oldated hashes, the cli
is going to show a warning with the necessary hashes to update and the
new ones.

Signed-off-by: Nathan Martins <nathan.martins@zup.com.br>