analyzer:bugfix - separate warnings from errors
Both errors and warnings are handled using the same field, this causes a
lot of confusion and a very complex code.
This pull request removes warnings from the errors field, and adds them
to the specific field for warnings. In the future with the refactoring
of formaters this code can be removed, but for now it is a necessary
workaround.

Signed-off-by: Nathan Martins <nathan.martins@zup.com.br>