formatters:chore - normalize not found files warn message
Previously the warning messages of files that was not found to execute
an analysis was different between formatters, and also the prefix
{HORUSEC_CLI} was being added on the middle of the message and not the
start.

This commit normalize these messages to have the same content for all
formatters and the {HORUSEC_CLI} prefix was fixed by adding on the start
of the message like all other log messages. A strings.ReplaceAll was
also added to remove the `;` suffix on these warning messages since
the commit 345c7487 change these messages from errors to warnings (
strings.ReplaceAll already exists to remove `;`from error messages).

The warn message of gemfile.lock file not found was removed since it
was not being used by any formatter.

Signed-off-by: Matheus Alcantara <matheus.alcantara@zup.com.br>