formatters/ruby:chore - removing unnecessary error messages (#1024)
Previously if the project had ruby between programming languages but is
not a ruby on rails project or does not have a Gemfile.lock an error
would be reported for each scenario.

This pull request changes these error messages introduced in commits
cd839ef and 9245d7d to be considered just warnings, as there is no
need to point them out as errors, since it is only intended to
inform the user.

Signed-off-by: Nathan Martins <nathan.martins@zup.com.br>
(cherry picked from commit 358fd4a5cd42118c442241155726a5dc385e0181)