goreleaser:bugfix - adding release auto
Before every release was released as a draft, in this way it was
possible to define which releases should be published as a pre release
and which not. After pull request 578e883, all releases are published
directly without the possibility to define betas and rcs as pre
releases.

This pull request changes the goreleaser configuration file to
automatically define which ones should be published as pre releases.
For this the pre release:auto attribute was added, for more information
follow the documentation link
https://goreleaser.com/customization/release/?h=release#release.

Signed-off-by: Nathan Martins <nathan.martins@zup.com.br>