workflow:feature - updating release workflows to semantic analysis
Previously it was not possible to build the binary for mac, as we did
not have a C compiler with a valid license in the main workflow. This
pull request adds a new workflow that is activated after completion of
the final-release, beta-release, or rc-release. This new workflow will
build the binaries directly on a mac machine and update the release with
the new binaries

Signed-off-by: Nathan Martins <nathan.martins@zup.com.br>