rules:feat - adding rule to spring framework rce
This commit adds a new rule to identify a new remote code execution
vulnerability in the spring framework. Due to the limitations of the
regex engine, this rule can bring some false positives about safe
versions pointed out as vulnerabilities. The rule will consider
any vulnerability < 5.3.18 as vulnerable, which is not true, as
versions >= 5.2.20 already have the fix for the problem, but due
to the limitation of the engine we can't detect it.

Signed-off-by: Nathan Martins <nathan.martins@zup.com.br>